reorganise tests - move edit test up to AccountUsers.test.tsx
