+++ static/js/src/advantage/users/AccountUsers.test.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import { render, screen, within, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
+import { User } from "./types";
 import AccountUsers from "./AccountUsers";
 import { mockData } from "./mockData";
 
@@ -29,4 +30,43 @@ it("displays a success message after adding a user", async () => {
   expect(screen.getByRole("alert")).toHaveTextContent(
     /User added successfully/
   );
+});
+
+it("allows to edit only a single user at a time", async () => {
+  const mockUserBase = {
+    lastLoginAt: "2021-06-10T09:05:00Z",
+  };
+  const users: User[] = [
+    {
+      ...mockUserBase,
+      id: "1",
+      name: "Karen",
+      email: "karen@ecorp.com",
+      role: "billing",
+    },
+    {
+      ...mockUserBase,
+      id: "3",
+      name: "Angela",
+      email: "angela@ecorp.com",
+      role: "technical",
+    },
+  ];
+
+  render(<AccountUsers organisationName="ECorp" users={users} />);
+
+  const EDIT_KAREN = "Edit user karen@ecorp.com";
+  const EDIT_ANGELA = "Edit user angela@ecorp.com";
+
+  screen
+    .getAllByRole("button", { name: /Edit/ })
+    .forEach((button) => expect(button).toBeEnabled());
+
+  userEvent.click(screen.getByLabelText(EDIT_KAREN));
+  expect(screen.getByLabelText(EDIT_ANGELA)).toBeDisabled();
+
+  userEvent.click(screen.getByRole("button", { name: "Cancel" }));
+
+  await waitFor(() => expect(screen.getByLabelText(EDIT_ANGELA)).toBeEnabled());
+  expect(screen.getByLabelText(EDIT_KAREN)).toBeEnabled();
 });
