+++ static/js/src/advantage/users/components/TableView/TableView.test.tsx
@@ -1,6 +1,5 @@
 import React from "react";
-import { render, screen, waitFor } from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
+import { render, screen } from "@testing-library/react";
 
 import { User } from "../../types";
 import TableView from "./TableView";
@@ -8,42 +7,24 @@ import TableView from "./TableView";
 it("displays user details in a correct format", () => {
   const testUser: User = {
     id: "1",
+    name: "User",
     email: "user@ecorp.com",
     role: "admin",
     lastLoginAt: "2021-02-15T13:45:00Z",
   };
 
-  render(<TableView users={[testUser]} />);
+  render(
+    <TableView
+      users={[testUser]}
+      userInEditModeById={null}
+      setUserInEditModeById={jest.fn()}
+      dismissEditMode={jest.fn()}
+      handleDeleteConfirmationModalOpen={jest.fn()}
+    />
+  );
 
   expect(screen.getByText("user@ecorp.com")).toBeInTheDocument();
   expect(screen.getByText("Admin", { ignore: "option" })).toBeInTheDocument();
   expect(screen.getByText("Admin", { selector: "option" })).not.toBeVisible();
   expect(screen.getByText("15/02/2021")).toBeInTheDocument();
-});
-
-it("allows to edit only a single user at a time", async () => {
-  const mockUserBase = {
-    lastLoginAt: "2021-06-10T09:05:00Z",
-  };
-  const users: User[] = [
-    { ...mockUserBase, id: "1", email: "karen@ecorp.com", role: "billing" },
-    { ...mockUserBase, id: "3", email: "angela@ecorp.com", role: "technical" },
-  ];
-
-  render(<TableView users={users} />);
-
-  const EDIT_KAREN = "Edit user karen@ecorp.com";
-  const EDIT_ANGELA = "Edit user angela@ecorp.com";
-
-  screen
-    .getAllByRole("button", { name: /Edit/ })
-    .forEach((button) => expect(button).toBeEnabled());
-
-  userEvent.click(screen.getByLabelText(EDIT_KAREN));
-  expect(screen.getByLabelText(EDIT_ANGELA)).toBeDisabled();
-
-  userEvent.click(screen.getByRole("button", { name: "Cancel" }));
-
-  await waitFor(() => expect(screen.getByLabelText(EDIT_ANGELA)).toBeEnabled());
-  expect(screen.getByLabelText(EDIT_KAREN)).toBeEnabled();
 });
